feat(declaration-responsive-page): WIP
