+++ front/src/components/common/Tooltip/index.ts
@@ -1,0 +1,1 @@
+export * from './Tooltip';
