+++ front/src/components/input/StandardInputs/Comboboxes/Comboboxes.tsx
@@ -62,7 +62,7 @@ export const Comboboxes: React.FC<ComboboxesOptions> = ({
 
   const className = classNames(fullWidth ? 'w-full' : 'max-w-fit');
   let classNameCombobox =
-    'w-full border py-2 pl-3 pr-10 focus:outline-none rounded-full placeholder:italic placeholder:text-secondary-400 placeholder:font-light';
+    'w-full border py-2 pl-3 pr-10 focus:outline-none rounded-full placeholder:italic placeholder:text-placeholder placeholder:font-light';
   classNameCombobox += error
     ? ' border-red-500 focus:ring-red-500 focus:border-red-500'
     : ' border-secondary-300';
@@ -93,7 +93,7 @@ export const Comboboxes: React.FC<ComboboxesOptions> = ({
           </div>
         )}
         {filteredOptions.length > 0 && query.length > 0 && (
-          <div className="absolute z-10 mt-1 max-h-60 w-full overflow-auto rounded-md bg-white p-0 py-1 text-base">
+          <div className="absolute z-10 mt-1 max-h-60 w-full overflow-auto rounded-md bg-white p-0 py-1 text-base md:text-xs">
             {filteredOptions.map((option) => (
               <div
                 key={option.id}
