+++ front/src/components/input/StandardInputs/Input/Input.tsx
@@ -68,7 +68,7 @@ export const Input: React.FC<IInputOptions> = ({
   }
   const className = twMerge(
     classNames({
-      'border-0 focus:outline-none focus:ring-0 text-base block w-full px-5 py-2 border-solid rounded-full focus:placeholder-transparent focus:outline-none focus:ring-transparent placeholder:italic placeholder:text-secondary-500 placeholder:font-light':
+      'border-0 focus:outline-none focus:ring-0 text-base md:text-xs block w-full px-5 py-2 md:py-[11px] border-solid rounded-full focus:placeholder-transparent focus:outline-none focus:ring-transparent placeholder:italic placeholder:text-placeholder placeholder:font-light':
         true,
       'border border-secondary-300 focus:border-secondary-300 ': withBorder,
       'text-error border border-red-500 focus:border-red-500 placeholder:text-red-300':
