+++ front/src/templates/DeclarationAgentStepsDesktop.tsx
@@ -3,8 +3,9 @@ import { ReactNode } from 'react';
 import classNames from 'classnames';
 import { UseFormHandleSubmit } from 'react-hook-form';
 
-import { MainAgent } from './MainAgent';
+import { MainAgentDesktop } from './MainAgentDesktop';
 import { ProgressBarAgent } from '@/components/common/ProgressBarAgent';
+import { ProgressBarAgentItemType } from '@/components/common/ProgressBarAgent/ProgressBarAgentItem';
 import { Meta } from '@/layout/Meta';
 import { DECLARATION_STEP_PAGE } from '@/utils/const';
 
@@ -17,58 +18,46 @@ type IMainProps = {
   linkButton?: string;
 };
 
-export type ProgressBarAgentItemType = {
-  name: string;
-  to: string;
-  stepNumber: number;
-};
-
 const links: ProgressBarAgentItemType[] = [
   { name: 'Coordonnées', to: DECLARATION_STEP_PAGE[1] ?? '', stepNumber: 1 },
   { name: 'Transports', to: DECLARATION_STEP_PAGE[2] ?? '', stepNumber: 2 },
   { name: 'Marchandises', to: DECLARATION_STEP_PAGE[3] ?? '', stepNumber: 3 },
 ];
 
-const DeclarationAgentSteps = ({
+const DeclarationAgentStepsDesktop = ({
   children,
   currentStep,
   handleSubmit,
   onSubmit,
   simpleBg,
-  linkButton,
 }: IMainProps) => {
   const showProgressBar = currentStep <= 3 && currentStep >= 1;
 
   return (
-    <MainAgent
-      meta={
-        <Meta
-          title="Simulateur Déclare Douanes"
-          description="Simuler la déclaration de douane en quelques clics"
-        />
-      }
-      withTitle
-      titleHeader="Créer une déclaration"
-      linkButton={linkButton}
-    >
-      {showProgressBar && (
-        <div className="px-4">
-          <ProgressBarAgent links={links} currentStep={currentStep} />
-        </div>
-      )}
-
+    <>
+      <MainAgentDesktop
+        meta={
+          <Meta
+            title="Simulateur Déclare Douanes"
+            description="Simuler la déclaration de douane en quelques clics"
+          />
+        }
+        titleHeader="Créer une déclaration"
+      >
+        {showProgressBar && <ProgressBarAgent links={links} currentStep={currentStep} />}
+      </MainAgentDesktop>
       <form
         onSubmit={handleSubmit(onSubmit)}
         className={classNames({
-          'flex flex-1 flex-col h-full p-5 mt-4': true,
+          'px-[126px]': true,
           'bg-white': simpleBg,
           'bg-secondary-bg': !simpleBg,
         })}
       >
         {children}
       </form>
-    </MainAgent>
+    </>
   );
 };
 
-export { DeclarationAgentSteps };
+export { DeclarationAgentStepsDesktop };
